% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paternity.R
\name{paternity}
\alias{paternity}
\title{Paternity by Exclusion Test}
\usage{
paternity(x)
}
\arguments{
\item{x}{matrix of values to test. The matrix should be 3 columns with the progeny being tested in the second column.}
}
\value{
A vector with proportion of cases that has a mismatch a.k.a. pedigree conflict.
}
\description{
Performs paternity by exclusion based on a matrix three entries (trio: mom, individual, dad), returns the proportion of mismatch cases.
}
\examples{
data(potato.data)
trio1 = potato.data[,c("W6511.1R", "W15268.5R", "VillettaRose")]
trio2 = potato.data[,c("W6511.1R", "W15268.5R", "W9914.1R")]
paternity(trio1) #more likely to be this trio
paternity(trio2)

}
