% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.data-data.R
\docType{data}
\name{potato.data}
\alias{potato.data}
\title{Potato data}
\format{
data.frame
}
\usage{
data(potato.data)
}
\description{
Potato data
}
\examples{
data(potato.data)

}
\keyword{datasets}
