% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Test parentage of individual}
\usage{
proportion(parents, individual, data)
}
\arguments{
\item{parents}{a vector with strings related to the name of the suspected parents}

\item{individual}{a string value with the individual name you are testing}

\item{data}{the dataframe from which the data is from}
}
\value{
A dataframe of different combinations of parents and individual with the proportion of pedigree conflicts in each trio
}
\description{
Given individual and a vectors of possible parents, function returns dataframe of proportion of pedigree conflict with each possible trio
}
\examples{
data(potato.data)
proportion(parents = c("W6511.1R","VillettaRose","W9914.1R"),
           individual = "W15268.1R",
           data = potato.data)

}
